requirements.txt
amqp==2.4.2
aniso8601==6.0.0
api2pdf==0.0.11
awscli==1.16.136
Babel==2.6.0
billiard==3.6.0.0
boto==2.49.0
boto3==1.09.90
botocore==1.12.126
celery==4.3.0
certifi==2018.8.13
chardet==3.0.4
Click==7.0
colorama==0.3.9
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
dal-admin-filters==0.3.3
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1
django-admin-autocomplete-filter==0.6.1
django-admin-list-filter-dropdown==1.0.2
django-admin-numeric-filter==0.1.3
django-admin-rangefilter==0.3.15
django-admin-sortable==2.1.8
django-admin-toolbox==1.0.0.dev26
django-ajax-selects==1.9.1
django-allauth==0.38.0
django-appconf==1.0.2
django-autocomplete-light==3.3.2
django-autoslug-iplweb==1.9.4
django-barcode==0.1.1
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-celerybeat-status==0.0.10
django-cors-headers==3.3.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-daterange-filter==1.3.0
django-debug-toolbar==1.11
django-elasticsearch-dsl==7.0.0
django-fcm==0.1.1
django-filter==2.0.0
django-graphiql==0.4.4
django-import-export==1.1.0
django-jet==1.0.8
django-js-asset==1.1.0
django-material==1.4.3
django-mathfilters==1.0.0
django-mptt==0.9.1
django-multidb-router==0.9
django-nested-admin==3.3.2
django-otp==0.5.0
django-redis==4.10.0
django-registration==3.0
django-rest-swagger==2.2.0
django-select2==6.3.1
django-sequences==2.4
django-ses==0.8.11
django-storages==1.7.1
django-tempus-dominus==5.0.1.4
django-timezone-field==3.0
django-wkhtmltopdf==3.3.0
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
drf-yasg==1.17.0
elastic-apm==5.3.2
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
et-xmlfile==1.0.1
Flask==1.0.3
future==0.17.1
google-api-python-client==1.4.1
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.7
importlib-metadata==1.6.1
inflect==2.1.0
inflection==0.5.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.3
Js2Py==0.70
jsonpickle==1.4.1
kombu==4.5.0
logstash==0.1.dev0
Markdown==2.6.11
MarkupSafe==1.1
num2words==0.5.10
oauth2client==4.1.3
oauthlib==2.1.0
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.6.2
packaging==20.4
Pillow~=7.2.0
promise==2.2.1
psutil==5.6.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyjsparser==2.7.1
pyexcel-xlsx==0.6.0
PyMySQL==0.9.3
pyotp==2.2.6
pyparsing==2.4.7
PyPDF2==1.26.0
python-barcode==0.9.0
python-crontab==2.3.6
python-dateutil==2.7.3
python-decouple==3.1
python-logstash==0.4.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
raven==6.10.0
redis==3.2.1
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Rx==1.6.1
s3transfer==0.2.0
sentry-sdk==0.7.4
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
stripe==2.10.1
tablib==0.12.1
tornado==6.0.2
tzlocal==2.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
uWSGI==2.0.17.1
vine==1.3.0
Werkzeug==0.15.4
wkhtmltopdf==0.2
xlrd==1.1.0
XlsxWriter==1.1.8
xlwt==1.3.0
zipp==3.1.0

django-model-utils~=4.0.0
pyodbc==4.0.30
django-tables2==2.3.4
django-multiselectfield==0.1.12